+++ packages/terra-core-docs/src/terra-dev-site/doc/toggle-button/example/OpenCloseEventToggleButton.jsx
@@ -29,8 +29,8 @@ const OpenCloseEventToggleButton = () => {
           <span className={cx('times-wrapper')}>{timesClosed}</span>
         </p>
       </div>
-      <ToggleButton closedButtonText="ToggleButton" onOpen={handleOnOpen} onClose={handleOnClose}>
-        <p>Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.</p>
+      <ToggleButton closedButtonText="Information about Model Experience" onOpen={handleOnOpen} onClose={handleOnClose}>
+        <p>New implementations of Cerner solutions use Model Experience, our collection of industry and Cerner best practices, to provide strategic recommendations for implementing and utilizing Cerner platforms. Cerner consulting services specialize in organizational change management and will help lead your organization through the implementation process. We’ll work with your employees to establish engagement and encourage best practices so your staff can confidently and competently adopt your new EHR system.</p>
       </ToggleButton>
     </div>
   );